requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
httplib2==0.20.4
lxml==4.7.1
oauth2client==4.1.3
oauthlib==3.2.0
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.0
webdriver_manager==3.5.2